+++ projects/plugins/jetpack/class.jetpack.php
@@ -16,6 +16,7 @@ use Automattic\Jetpack\Constants;
 use Automattic\Jetpack\Device_Detection\User_Agent_Info;
 use Automattic\Jetpack\Identity_Crisis;
 use Automattic\Jetpack\Licensing;
+use Automattic\Jetpack\My_Jetpack\Initializer as My_Jetpack_Initializer;
 use Automattic\Jetpack\Partner;
 use Automattic\Jetpack\Plugin\Tracking as Plugin_Tracking;
 use Automattic\Jetpack\Redirect;
@@ -845,18 +846,6 @@ class Jetpack {
 
 		if ( ! $this->connection_manager ) {
 			$this->connection_manager = new Connection_Manager( 'jetpack' );
-
-			/**
-			 * Filter to activate Jetpack Connection UI.
-			 * INTERNAL USE ONLY.
-			 *
-			 * @since 9.5.0
-			 *
-			 * @param bool false Whether to activate the Connection UI.
-			 */
-			if ( apply_filters( 'jetpack_connection_ui_active', false ) ) {
-				Automattic\Jetpack\ConnectionUI\Admin::init();
-			}
 		}
 
 		/*
@@ -914,6 +903,7 @@ class Jetpack {
 		add_action( 'plugins_loaded', array( 'Jetpack', 'load_modules' ), 100 );
 
 		Partner::init();
+		My_Jetpack_Initializer::init();
 
 		/**
 		 * Fires when Jetpack is fully loaded and ready. This is the point where it's safe
@@ -6449,7 +6439,7 @@ endif;
 					height: 20px;
 				}
 				.jp-emblem-user-admin path {
-					fill: #00BE28;
+					fill: #069e08;
 				}
 			</style>
 			<?php
@@ -6705,11 +6695,10 @@ endif;
 			'discount_percent'  => 40,
 			'included_in_plans' => array( 'security' ),
 			'features'          => array(
+				_x( 'Real-time cloud backups', 'Backup Product Feature', 'jetpack' ),
 				_x( '10GB of backup storage', 'Backup Product Feature', 'jetpack' ),
-				_x( 'One-click restore from the last 30 days of backups', 'Backup Product Feature', 'jetpack' ),
-				_x( '30-day activity log', 'Backup Product Feature', 'jetpack' ),
-				_x( 'Real-time backups (as you edit)', 'Backup Product Feature', 'jetpack' ),
-				_x( 'Cloud-based, secure backups', 'Backup Product Feature', 'jetpack' ),
+				_x( '30-day archive & activity log', 'Backup Product Feature', 'jetpack' ),
+				_x( 'One-click restores', 'Backup Product Feature', 'jetpack' ),
 			),
 		);
 
@@ -6758,14 +6747,14 @@ endif;
 		);
 
 		$products['security'] = array(
-			'title'             => __( 'Security Bundle', 'jetpack' ),
+			'title'             => __( 'Security', 'jetpack' ),
 			'slug'              => 'jetpack_security_t1_yearly',
-			'description'       => __( 'Get all security products including backups, site scanning, and anti-spam.', 'jetpack' ),
+			'description'       => __( 'Comprehensive site security, including Backup, Scan, and Anti-spam.', 'jetpack' ),
 			'show_promotion'    => true,
 			'discount_percent'  => 40,
 			'included_in_plans' => array(),
 			'features'          => array(
-				_x( 'All Backup Features ', 'Security Daily Plan Feature', 'jetpack' ),
+				_x( 'Real-time cloud backups with 10GB storage', 'Security Tier 1 Feature', 'jetpack' ),
 				_x( 'Automated real-time malware scan', 'Security Daily Plan Feature', 'jetpack' ),
 				_x( 'One-click fixes for most threats', 'Security Daily Plan Feature', 'jetpack' ),
 				_x( 'Comment & form spam protection', 'Security Daily Plan Feature', 'jetpack' ),